requirements.txt
langchain
openai
langchain_community
huggingface_hub
python-dotenv
streamlit